Move test files to mirror the application file structure.
