+++ tests/conftest.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm.session import Session
 from slackhealthbot.data.database.models import Base
 from slackhealthbot.main import app
 from slackhealthbot.routers.dependencies import get_db
-from tests.factories.factories import (
+from tests.testsupport.factories.factories import (
     FitbitActivityFactory,
     FitbitUserFactory,
     UserFactory,
