+++ tests/testsupport/fixtures/fitbit_scenarios.py
