+++ tests/testsupport/fixtures/fitbit_sleep_response_1_item.json
