+++ tests/testsupport/fixtures/fitbit_sleep_response_classic.json
