+++ tests/testsupport/fixtures/fitbit_sleep_response_no_main_sleep_item.json
